merge with upstream repeate

+++ crates/revm/src/db/states/state.rs
@@ -171,6 +171,9 @@ impl<DB: Database> State<DB> {
         }
     }
 
+    /// Get a mutable reference to the [`CacheAccount`] for the given address.
+    /// If the account is not found in the cache, it will be loaded from the
+    /// database and inserted into the cache.
     pub fn load_cache_account(&mut self, address: Address) -> Result<&mut CacheAccount, DB::Error> {
         match self.cache.accounts.entry(address) {
             hash_map::Entry::Vacant(entry) => {
@@ -198,15 +201,15 @@ impl<DB: Database> State<DB> {
     }
 
     // TODO make cache aware of transitions dropping by having global transition counter.
-    /// Takes changeset and reverts from state and replaces it with empty one.
-    /// This will trop pending Transition and any transitions would be lost.
+    /// Takes the [`BundleState`] changeset from the [`State`], replacing it
+    /// with an empty one.
     ///
-    /// NOTE: If either:
-    /// * The [State] has not been built with [StateBuilder::with_bundle_update], or
-    /// * The [State] has a [TransitionState] set to `None` when
-    ///   [State::merge_transitions] is called,
+    /// This will not apply any pending [`TransitionState`]. It is recommended
+    /// to call [`State::merge_transitions`] before taking the bundle.
     ///
-    /// this will panic.
+    /// If the `State` has been built with the
+    /// [`StateBuilder::with_bundle_prestate`] option, the pre-state will be
+    /// taken along with any changes made by [`State::merge_transitions`].
     pub fn take_bundle(&mut self) -> BundleState {
         core::mem::take(&mut self.bundle_state)
     }
@@ -275,7 +278,7 @@ impl<DB: Database> Database for State<DB> {
                 let ret = *entry.insert(self.database.block_hash(number)?);
 
                 // prune all hashes that are older then BLOCK_HASH_HISTORY
-                let last_block = number.saturating_sub(BLOCK_HASH_HISTORY as u64);
+                let last_block = number.saturating_sub(BLOCK_HASH_HISTORY);
                 while let Some(entry) = self.block_hashes.first_entry() {
                     if *entry.key() < last_block {
                         entry.remove();
@@ -312,7 +315,7 @@ mod tests {
         state.block_hash(1u64).unwrap();
         state.block_hash(2u64).unwrap();
 
-        let test_number = BLOCK_HASH_HISTORY as u64 + 2;
+        let test_number = BLOCK_HASH_HISTORY + 2;
 
         let block1_hash = keccak256(U256::from(1).to_string().as_bytes());
         let block2_hash = keccak256(U256::from(2).to_string().as_bytes());
